+++ lib/Braintree/Customer.php
@@ -456,19 +456,6 @@ class Braintree_Customer extends Braintree
     }
 
     /**
-     * sets private properties
-     * this function is private so values are read only
-     * @ignore
-     * @access protected
-     * @param string $key
-     * @param mixed $value
-     */
-    private function _set($key, $value)
-    {
-        $this->_attributes[$key] = $value;
-    }
-
-    /**
      * verifies that a valid customer id is being used
      * @ignore
      * @param string customer id