+++ tests/unit/TransactionTest.php
@@ -6,7 +6,7 @@ class Braintree_TransactionTest extends PHPUnit_Framework_TestCase
     function testGet_givesErrorIfInvalidProperty()
     {
         $t = Braintree_Transaction::factory(array(
-            'creditCard' => array(),
+            'creditCard' => array('expirationMonth' => '05', 'expirationYear' => '2010', 'bin' => '510510', 'last4' => '5100'),
             'customer' => array(),
             'billing' => array(),
             'shipping' => array(),